Refined the Mercurial command server interface.

+++ Plugins/VcsPlugins/vcsMercurial/BookmarksExtension/HgBookmarksListDialog.py
@@ -55,11 +55,15 @@ class HgBookmarksListDialog(QDialog, Ui_HgBookmarksListDialog):
         
         @param e close event (QCloseEvent)
         """
-        if self.process is not None and \
-           self.process.state() != QProcess.NotRunning:
-            self.process.terminate()
-            QTimer.singleShot(2000, self.process.kill)
-            self.process.waitForFinished(3000)
+        if self.__hgClient:
+            if self.__hgClient.isExecuting():
+                self.__hgClient.cancel()
+        else:
+            if self.process is not None and \
+               self.process.state() != QProcess.NotRunning:
+                self.process.terminate()
+                QTimer.singleShot(2000, self.process.kill)
+                self.process.waitForFinished(3000)
         
         e.accept()
     
@@ -99,6 +103,8 @@ class HgBookmarksListDialog(QDialog, Ui_HgBookmarksListDialog):
             if out:
                 for line in out.splitlines():
                     self.__processOutputLine(line)
+                    if self.__hgClient.wasCanceled():
+                        break
             self.__finish()
         else:
             self.process.kill()
