+++ Plugins/VcsPlugins/vcsMercurial/GpgExtension/HgGpgSignaturesDialog.py
@@ -54,11 +54,15 @@ class HgGpgSignaturesDialog(QDialog, Ui_HgGpgSignaturesDialog):
         
         @param e close event (QCloseEvent)
         """
-        if self.process is not None and \
-           self.process.state() != QProcess.NotRunning:
-            self.process.terminate()
-            QTimer.singleShot(2000, self.process.kill)
-            self.process.waitForFinished(3000)
+        if self.__hgClient:
+            if self.__hgClient.isExecuting():
+                self.__hgClient.cancel()
+        else:
+            if self.process is not None and \
+               self.process.state() != QProcess.NotRunning:
+                self.process.terminate()
+                QTimer.singleShot(2000, self.process.kill)
+                self.process.waitForFinished(3000)
         
         e.accept()
     
@@ -96,6 +100,8 @@ class HgGpgSignaturesDialog(QDialog, Ui_HgGpgSignaturesDialog):
             if out:
                 for line in out.splitlines():
                     self.__processOutputLine(line)
+                    if self.__hgClient.wasCanceled():
+                        break
             self.__finish()
         else:
             self.process.kill()
